splitting mcmc tests into separate modules per class
